separate fixtures and functions to dedicated files
